Cleanup redundant option-parsing code.
Recently, code was added to fixpatch.py and exportpatch.py
that augmented the option parsing class so that it did not
call sys.exit(() when encountering an error.

This code was duplicated in fixpatch.py and exportpatch.py,
so this commit creates a new file with the common code,
so that the existing modules can just import the functionality.
