+++ patchtools/exportpatch.py
@@ -8,8 +8,8 @@ __author__ = 'Jeff Mahoney'
 
 import sys
 from patchtools import PatchException
+from patchtools.modified_optparse import ModifiedOptionParser, OptionParsingError
 from patchtools.patch import Patch, EmptyCommitException
-from optparse import OptionParser
 import os
 
 
@@ -64,23 +64,6 @@ def export_patch(commit, options, prefix, suffix):
 
     print("Couldn't locate commit \"%s\"; Skipping." % commit, file=sys.stderr)
     return 1
-
-#
-# set up Option Parsing class so that we can
-# stop the option parser from calling sys.exit()
-# when it encounters an error
-#
-
-class OptionParsingError(RuntimeError):
-    """An exception raised when parser.error() is called."""
-    def __init__(self, msg):
-        self.msg = msg
-
-
-class ModifiedOptionParser(OptionParser):
-    """Our own Option Parsing class, that does not call sys.exit()."""
-    def error(self, msg):
-        raise OptionParsingError(msg)
 
 
 def main():
